# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Atoma API
  description: |
    REST API for Atoma - a minimalist iOS app for habit tracking, task management, and budget control.

    **Tagline**: habits, tasks, money - in one flow

    ## Authentication
    All endpoints (except auth) require a valid JWT access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Response Format
    All successful responses follow this format:
    ```json
    { "data": { ... } }
    ```

    Error responses:
    ```json
    { "error": { "code": "ERROR_CODE", "message": "..." } }
    ```
  version: 1.0.0
  contact:
    name: Atoma Support
    url: https://api.azamatbigali.online
  license:
    name: Private
servers:
  - url: https://api.azamatbigali.online/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /auth/google:
    post:
      tags: [Auth]
      summary: Google Sign-In
      description: Authenticate with Google ID token
      operationId: googleSignIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleSignInRequest'
      responses:
        '200':
          description: Existing user authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '201':
          description: New user created and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/apple:
    post:
      tags: [Auth]
      summary: Apple Sign-In
      description: Authenticate with Apple ID token
      operationId: appleSignIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppleSignInRequest'
      responses:
        '200':
          description: Existing user authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '201':
          description: New user created and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh Token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get Current User
      description: Get authenticated user profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      description: Invalidate refresh token
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/account:
    delete:
      tags: [Auth]
      summary: Delete Account
      description: Permanently delete user account and all data
      operationId: deleteAccount
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== HABITS ====================
  /habits:
    get:
      tags: [Habits]
      summary: List Habits
      description: Get all habits for the authenticated user
      operationId: listHabits
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of habits
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Habit'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Habits]
      summary: Create Habit
      description: Create a new habit
      operationId: createHabit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHabitRequest'
      responses:
        '201':
          description: Habit created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Habit'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /habits/{id}:
    get:
      tags: [Habits]
      summary: Get Habit
      description: Get a single habit by ID
      operationId: getHabit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Habit details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Habit'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Habits]
      summary: Update Habit
      description: Update an existing habit
      operationId: updateHabit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHabitRequest'
      responses:
        '200':
          description: Habit updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Habit'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Habits]
      summary: Delete Habit
      description: Delete a habit
      operationId: deleteHabit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Habit deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /habits/{id}/toggle:
    post:
      tags: [Habits]
      summary: Toggle Completion
      description: Toggle habit completion for a specific date
      operationId: toggleHabitCompletion
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToggleCompletionRequest'
      responses:
        '200':
          description: Completion toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Habit'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== TASKS ====================
  /tasks:
    get:
      tags: [Tasks]
      summary: List Tasks
      description: Get all tasks, optionally filtered by date
      operationId: listTasks
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          description: Filter by date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTasksResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Tasks]
      summary: Create Task
      description: Create a new task
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get Task
      description: Get a single task by ID
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tasks]
      summary: Update Task
      description: Update an existing task
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tasks]
      summary: Delete Task
      description: Delete a task
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{id}/toggle:
    post:
      tags: [Tasks]
      summary: Toggle Task
      description: Toggle task completion status
      operationId: toggleTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Task toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== TRANSACTIONS ====================
  /transactions:
    get:
      tags: [Transactions]
      summary: List Transactions
      description: Get all transactions, optionally filtered by month
      operationId: listTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          description: Filter by year
          schema:
            type: integer
            example: 2024
        - name: month
          in: query
          description: Filter by month (1-12)
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactionsResponse'

    post:
      tags: [Transactions]
      summary: Create Transaction
      description: Create a new income or expense transaction
      operationId: createTransaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/ValidationError'
  /transactions/summary:
    get:
      tags: [Transactions]
      summary: Get Summary
      description: Get income/expense summary for a month
      operationId: getTransactionSummary
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          description: Year (defaults to current)
          schema:
            type: integer
        - name: month
          in: query
          description: Month 1-12 (defaults to current)
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        '200':
          description: Monthly summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SummaryResponse'
  /transactions/{id}:
    get:
      tags: [Transactions]
      summary: Get Transaction
      description: Get a single transaction by ID
      operationId: getTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Transactions]
      summary: Update Transaction
      description: Update an existing transaction
      operationId: updateTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionRequest'
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Transactions]
      summary: Delete Transaction
      description: Delete a transaction
      operationId: deleteTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Transaction deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== RECURRING TRANSACTIONS ====================
  /recurring-transactions:
    get:
      tags: [Recurring Transactions]
      summary: List Recurring Transactions
      description: Get all recurring transactions
      operationId: listRecurringTransactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of recurring transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecurringTransaction'

    post:
      tags: [Recurring Transactions]
      summary: Create Recurring Transaction
      description: Create a new recurring transaction
      operationId: createRecurringTransaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecurringRequest'
      responses:
        '201':
          description: Recurring transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RecurringTransaction'
        '400':
          $ref: '#/components/responses/ValidationError'
  /recurring-transactions/projection:
    get:
      tags: [Recurring Transactions]
      summary: Get Projection
      description: Get monthly projection for recurring transactions
      operationId: getRecurringProjection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Monthly projection
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ProjectionResponse'
  /recurring-transactions/process:
    post:
      tags: [Recurring Transactions]
      summary: Process Recurring
      description: Process due recurring transactions and create actual transactions
      operationId: processRecurringTransactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Processing result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      processed:
                        type: integer
                      transactions_created:
                        type: integer
  /recurring-transactions/{id}:
    get:
      tags: [Recurring Transactions]
      summary: Get Recurring Transaction
      operationId: getRecurringTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Recurring transaction details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RecurringTransaction'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Recurring Transactions]
      summary: Update Recurring Transaction
      operationId: updateRecurringTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecurringRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RecurringTransaction'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Recurring Transactions]
      summary: Delete Recurring Transaction
      operationId: deleteRecurringTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== GOALS ====================
  /goals:
    get:
      tags: [Goals]
      summary: List Goals
      description: Get all long-term goals
      operationId: listGoals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of goals
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Goal'

    post:
      tags: [Goals]
      summary: Create Goal
      operationId: createGoal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Goal created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Goal'
  /goals/{id}:
    get:
      tags: [Goals]
      summary: Get Goal
      operationId: getGoal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Goal details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Goal'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Goals]
      summary: Update Goal
      operationId: updateGoal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalRequest'
      responses:
        '200':
          description: Goal updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Goal'

    delete:
      tags: [Goals]
      summary: Delete Goal
      operationId: deleteGoal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Goal deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # ==================== SAVINGS GOAL ====================
  /savings-goal:
    get:
      tags: [Savings]
      summary: Get Savings Goal
      description: Get user's monthly savings goal with progress
      operationId: getSavingsGoal
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Savings goal with progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SavingsGoalWithProgress'

    post:
      tags: [Savings]
      summary: Set Savings Goal
      description: Create or update monthly savings goal
      operationId: setSavingsGoal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSavingsGoalRequest'
      responses:
        '200':
          description: Savings goal set
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SavingsGoalWithProgress'

    delete:
      tags: [Savings]
      summary: Delete Savings Goal
      operationId: deleteSavingsGoal
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Savings goal deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean

  # ==================== AI ====================
...
  /ai/insights:
    post:
      tags: [AI]
      summary: Generate Insights
      description: Generate AI insights for habits, tasks, budget, or weekly review
      operationId: generateInsights
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsightRequest'
      responses:
        '200':
          description: Generated insights
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/InsightResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...
  schemas:
...
    GoogleSignInRequest:
      type: object
      required: [id_token]
      properties:
        id_token:
          type: string
          description: Google ID token from Sign-In
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    AuthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tokens:
              $ref: '#/components/schemas/TokenPair'
            is_new_user:
              type: boolean
...